# docs/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: TRM Referral Platform API
  description: |
    # TRM Referral Platform API
    
    The TRM (Talent Referral Marketplace) API enables developers to build on the referral platform.
    Manage job postings, referrals, companies, and receive real-time webhook notifications.
    
    ## Base URL
    - Production: `https://api.trm.com/v1`
    - Staging: `https://api.staging.trm.com/v1`
    
    ## Authentication
    All API requests require authentication using an API key passed in the `X-API-Key` header.
    
    ## Rate Limiting
    - Standard tier: 100 requests per 15 minutes
    - Enterprise tier: 1000 requests per 15 minutes
    
    ## Myanmar Payment Support
    The platform supports Myanmar-specific payment methods:
    - KBZPay
    - WavePay
    - CB Pay
    - Bank Transfers (local Myanmar banks)
    
  version: 1.0.0
  contact:
    name: TRM API Support
    email: api-support@trm.com
    url: https://docs.trm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trm.com/v1
    description: Production server
  - url: https://api.staging.trm.com/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /companies:
    get:
      summary: List companies
      description: Retrieve a list of companies with optional filtering
      tags:
        - Companies
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
            enum: [pending, verified, rejected]
      responses:
        '200':
          description: Successful response

    post:
      summary: Create company
      description: Create a new company (requires admin or corporate_admin role)
      tags:
        - Companies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateRequest'
      responses:
        '201':
          description: Company created successfully
  /companies/{id}:
    get:
      summary: Get company
      description: Get detailed information about a company
      tags:
        - Companies
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'

    put:
      summary: Update company
      description: Update a company (requires companies:write permission)
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdateRequest'
      responses:
        '200':
          description: Company updated successfully
  /companies/{id}/jobs:
    get:
      summary: Get company jobs
      description: Get all jobs for a specific company
      tags:
        - Companies
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /companies/{id}/team:
    get:
      summary: Get company team
      description: Get all team members for a company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response

    post:
      summary: Invite team member
      description: Invite a new team member to the company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [admin, recruiter, viewer]
                name:
                  type: string
      responses:
        '201':
          description: Invitation sent successfully
  /companies/{id}/team/{userId}:
    delete:
      summary: Remove team member
      description: Remove a team member from the company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Team member removed successfully
  /companies/{id}/billing:
    get:
      summary: Get company billing
      description: Get billing history for a company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
  /companies/{id}/analytics:
    get:
      summary: Get company analytics
      description: Get analytics data for a company
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        '200':
          description: Successful response
  /companies/{id}/verify:
    put:
      summary: Verify company
      description: Verify a company (admin only)
      tags:
        - Companies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [verified, rejected]
                notes:
                  type: string
      responses:
        '200':
          description: Company verification updated

  # ==================== USERS ====================
...
  /admin/companies:
    get:
      summary: Get all companies
      description: Get all companies with optional filtering
      tags:
        - Admin
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication for programmatic access
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for user sessions
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        logo:
          type: string
        website:
          type: string
        industry:
          type: string
        companySize:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            country:
              type: string
        verificationStatus:
          type: string
          enum: [pending, verified, rejected]
        currentSubscription:
          type: object
          properties:
            planId:
              type: string
            status:
              type: string
            expiresAt:
              type: string
              format: date-time
        stats:
          type: object
          properties:
            totalJobsPosted:
              type: integer
            totalHires:
              type: integer
            totalReferralSpend:
              type: number
        createdAt:
          type: string
          format: date-time
    CompanyCreateRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        description:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
        industry:
          type: string
        companySize:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            city:
              type: string
            state:
              type: string
            country:
              type: string
    CompanyUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logo:
          type: string
        website:
          type: string
        industry:
          type: string
        companySize:
          type: string
        address:
          type: object

    # ==================== REFERRAL SCHEMAS ====================
...